( #5 , #6 , #9 ) : memory ipc, data format write 데이터 포맷 규칙 통일화, hubsender에서 redis 읽기 기능 추가     - agent os의 파이프를 :로 변경 및 카테고리 추가     - agent redis 시퀀스 타입 uint64로 변경     - agent redis tcp 기능 주석 처리 및 미실행으로 변경     - hub sender에 redis 읽기 기능 추가

+++ agent_redis/pkg/format/format.go
@@ -6,7 +6,7 @@ import (
 	"encoding/binary"
 )
 
-func MakeFormat(seq int, data []byte) ([]byte, error) {
+func MakeFormat(seq uint64, data []byte) ([]byte, error) {
 	temp := &bytes.Buffer{}
 	buf := bufio.NewWriter(temp)
 
@@ -16,7 +16,7 @@ func MakeFormat(seq int, data []byte) ([]byte, error) {
 		return nil, lErr
 	}
 	buf.WriteByte(':')
-	err := binary.Write(buf, binary.LittleEndian, uint32(seq))
+	err := binary.Write(buf, binary.LittleEndian, seq)
 	if err != nil {
 		return nil ,err
 	}
